+++ redux_pytest/fixtures/event_loop.py
@@ -12,15 +12,17 @@ class LoopThread(threading.Thread):
     def __init__(self: LoopThread) -> None:
         super().__init__()
         self.loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(self.loop)
 
     def run(self: LoopThread) -> None:
         self.loop.run_forever()
 
     def stop(self: LoopThread) -> None:
+        asyncio.set_event_loop(None)
         self.loop.call_soon_threadsafe(self.loop.stop)
 
     def create_task(self: LoopThread, coro: Coroutine) -> None:
-        self.loop.call_soon_threadsafe(lambda: self.loop.create_task(coro))
+        self.loop.call_soon_threadsafe(self.loop.create_task, coro)
 
 
 @pytest.fixture()