+++ tests/test_scheduler.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import asyncio
 import threading
 from dataclasses import replace
-from typing import TYPE_CHECKING, Callable, TypeAlias
+from typing import TYPE_CHECKING, Callable, Coroutine, TypeAlias
 from unittest.mock import call
 
 from immutable import Immutable
@@ -50,6 +50,7 @@ class Scheduler(threading.Thread):
         self.stopped = False
         self._callbacks: list[tuple[Callable[[], None], float]] = []
         self.loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(self.loop)
         self.tasks: set[asyncio.Task] = set()
 
     def run(self: Scheduler) -> None:
@@ -85,17 +86,27 @@ class Scheduler(threading.Thread):
         self.loop.call_soon_threadsafe(self.loop.create_task, self.graceful_stop())
 
 
-def test_scheduler(
-    mocker: MockerFixture,
-) -> None:
+def test_scheduler(mocker: MockerFixture) -> None:
     scheduler = Scheduler()
     scheduler.start()
+
+    def _create_task_with_callback(
+        coro: Coroutine,
+        callback: Callable[[asyncio.Task], None] | None = None,
+    ) -> None:
+        def create_task_with_callback() -> None:
+            task = scheduler.loop.create_task(coro)
+            if callback:
+                callback(task)
+
+        scheduler.loop.call_soon_threadsafe(create_task_with_callback)
 
     store = Store(
         reducer,
         options=CreateStoreOptions(
             auto_init=True,
             scheduler=scheduler.set,
+            task_creator=_create_task_with_callback,
             on_finish=scheduler.schedule_stop,
         ),
     )
@@ -114,8 +125,7 @@ def test_scheduler(
     scheduler.join()
 
     render.assert_has_calls(
-        [
-            call(StateType(value=0)),
-            call(StateType(value=10)),
-        ],
+        [call(StateType(value=i)) for i in range(11)] + [call(StateType(value=10))],
     )
+
+    print(3)